requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
install==1.3.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.14.0
python-Levenshtein==0.12.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
spotipy==2.16.0
uritemplate==3.0.1
urllib3==1.25.10
ytmusicapi==0.10.2
